alu.sv
This module implements an ALU module that takes in a 4-bit opcode, two 8 bit inputs, a clock signal, output enable signal and a reset signal.  At its output, four flags are set according to the arithmetic performed and not the operations.  They are overflow flag (OF), carry flag (CF), zero flag (ZF), and a negative flag (SF). Along with the flags, this ALU also outputs the arithmetic result according to the opcode.